+++ src/util.rs
@@ -76,10 +76,10 @@ pub fn classify(keyword: &str) -> KeywordClass {
 
 pub fn expand(x: &str, len: usize, align: &ConfigColumnAlign) -> String {
     match align {
-        ConfigColumnAlign::Left => format!("{}{}", x, " ".repeat(len - x.len())),
-        ConfigColumnAlign::Right => format!("{}{}", " ".repeat(len - x.len()), x),
+        ConfigColumnAlign::Left => format!("{}{}", x, " ".repeat(len - x.chars().count())),
+        ConfigColumnAlign::Right => format!("{}{}", " ".repeat(len - x.chars().count()), x),
         ConfigColumnAlign::Center => {
-            let space = len - x.len();
+            let space = len - x.chars().count();
             let left = space / 2;
             let right = space / 2 + space % 2;
             format!("{}{}{}", " ".repeat(left), x, " ".repeat(right))