+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.5.6...Unreleased) - ReleaseDate
 
+* [Fixed] unmatched `--sort*` affects sort order
+
 ## [v0.5.6](https://github.com/dalance/procs/compare/v0.5.5...v0.5.6) - 2019-03-01
 
 * [Added] Ssb column