+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [Added] separator option to `~/.procs.toml` setting
 * [Added] `--list` option to show column kind list
 * [Added] Slot column to insert column by `--insert` oprion
+* [Added] Sort indicator
 * [Changed] the first decimal place of day/year in CpuTime is shown
 * [Changed] default separator from "|" to "│" ( U+2502:Box Drawings Light Vertical )
 * [Changed] eip/esp/sig* format to 16 hex digits