+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-## [Unreleased](https://github.com/dalance/procs/compare/v0.5.6...Unreleased) - ReleaseDate
+## [Unreleased](https://github.com/dalance/procs/compare/v0.5.7...Unreleased) - ReleaseDate
+
+## [v0.5.7](https://github.com/dalance/procs/compare/v0.5.6...v0.5.7) - 2019-03-05
 
 * [Added] separator option to `~/.procs.toml` setting
 * [Added] `--list` option to show column kind list